PCHR-2364: Removed the DataImport Trait. Moved the methods to the Base parser file.
